+++ src/main/java/no/nsd/qddt/domain/category/web/CategoryAuditController.java
@@ -34,28 +34,28 @@ public class CategoryAuditController {
 
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-    public Revision<Integer, Category> getLastRevision(@PathVariable("id") UUID id) {
+    public Revision<Long, Category> getLastRevision(@PathVariable("id") UUID id) {
         return auditService.findLastChange(id);
     }
 
     @RequestMapping(value = "/{id}/{revision}", method = RequestMethod.GET)
-    public Revision<Integer, Category> getByRevision(@PathVariable("id") UUID id, @PathVariable("revision") Integer revision) {
+    public Revision<Long, Category> getByRevision(@PathVariable("id") UUID id, @PathVariable("revision") Long revision) {
         return auditService.findRevision(id, revision);
     }
 
     @RequestMapping(value = "/{id}/all", method = RequestMethod.GET)
-    public HttpEntity<PagedResources<Revision<Integer, Category>>> allProjects(
+    public HttpEntity<PagedResources<Revision<Long, Category>>> allProjects(
             @PathVariable("id") UUID id,
             @RequestParam(value = "ignorechangekinds",defaultValue = "IN_DEVELOPMENT,UPDATED_HIERARCHY_RELATION,UPDATED_PARENT")
             Collection<AbstractEntityAudit.ChangeKind> changekinds,
             Pageable pageable, PagedResourcesAssembler assembler) {
 
-        Page<Revision<Integer, Category>> entities = auditService.findRevisionByIdAndChangeKindNotIn(id,changekinds, pageable);
+        Page<Revision<Long, Category>> entities = auditService.findRevisionByIdAndChangeKindNotIn(id,changekinds, pageable);
         return new ResponseEntity<>(assembler.toResource(entities), HttpStatus.OK);
     }
 
 //    @RequestMapping(value = "/version/{id}/{version}", method = RequestMethod.GET)
-//    public Revision<Integer, Category> getByVersion(@PathVariable("id") UUID id, @PathVariable("version") String version) {
+//    public Revision<Long, Category> getByVersion(@PathVariable("id") UUID id, @PathVariable("version") String version) {
 //        return auditService.findVersion(id, version);
 //    }
