+++ src/main/java/no/nsd/qddt/domain/conceptquestionitem/ConceptQuestionItemJson.java
@@ -15,7 +15,7 @@ public class ConceptQuestionItemJson {
 
     private QuestionItemJsonView questionItem;
 
-    private Integer questionItemRevision;
+    private Long questionItemRevision;
 
 //    @JsonSerialize(using = JsonDateSerializer.class)
     private Timestamp updated;
@@ -48,11 +48,11 @@ public class ConceptQuestionItemJson {
         this.questionItem = questionItem;
     }
 
-    public Integer getQuestionItemRevision() {
+    public Long getQuestionItemRevision() {
         return questionItemRevision;
     }
 
-    private void setQuestionItemRevision(Integer questionItemRevision) {
+    private void setQuestionItemRevision(Long questionItemRevision) {
         this.questionItemRevision = questionItemRevision;
     }
