+++ src/main/java/no/nsd/qddt/domain/conceptquestionitem/ParentQuestionItem.java
@@ -6,6 +6,6 @@ package no.nsd.qddt.domain.conceptquestionitem;
 public interface ParentQuestionItem {
 
     ParentQuestionItemId getId();
-    Integer getQuestionItemRevision();
+    Long getQuestionItemRevision();
 
 }
