+++ src/main/java/no/nsd/qddt/domain/instruction/audit/InstructionAuditService.java
@@ -13,7 +13,7 @@ import java.util.UUID;
 /**
  * @author Dag Østgulen Heradstveit
  */
-public interface InstructionAuditService extends BaseServiceAudit<Instruction, UUID, Integer> {
+public interface InstructionAuditService extends BaseServiceAudit<Instruction, UUID, Long> {
 
-    Page<Revision<Integer, Instruction>> findRevisionByIdAndChangeKindNotIn(UUID id, Collection<AbstractEntityAudit.ChangeKind> changeKinds, Pageable pageable);
+    Page<Revision<Long, Instruction>> findRevisionByIdAndChangeKindNotIn(UUID id, Collection<AbstractEntityAudit.ChangeKind> changeKinds, Pageable pageable);
 }
