+++ src/main/java/no/nsd/qddt/domain/instruction/web/InstructionAuditController.java
@@ -34,23 +34,23 @@ public class InstructionAuditController {
 
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-    public Revision<Integer, Instruction> getLastRevision(@PathVariable("id") UUID id) {
+    public Revision<Long, Instruction> getLastRevision(@PathVariable("id") UUID id) {
         return auditService.findLastChange(id);
     }
 
     @RequestMapping(value = "/{id}/{revision}", method = RequestMethod.GET)
-    public Revision<Integer, Instruction> getByRevision(@PathVariable("id") UUID id, @PathVariable("revision") Integer revision) {
+    public Revision<Long, Instruction> getByRevision(@PathVariable("id") UUID id, @PathVariable("revision") Long revision) {
         return auditService.findRevision(id, revision);
     }
 
     @RequestMapping(value = "/{id}/all", method = RequestMethod.GET)
-    public HttpEntity<PagedResources<Revision<Integer, Instruction>>> allProjects(
+    public HttpEntity<PagedResources<Revision<Long, Instruction>>> allProjects(
             @PathVariable("id") UUID id,
             @RequestParam(value = "ignorechangekinds",defaultValue = "IN_DEVELOPMENT,UPDATED_HIERARCHY_RELATION,UPDATED_PARENT")
                     Collection<AbstractEntityAudit.ChangeKind> changekinds,
             Pageable pageable, PagedResourcesAssembler assembler) {
 
-        Page<Revision<Integer, Instruction>> entities = auditService.findRevisionByIdAndChangeKindNotIn(id,changekinds, pageable);
+        Page<Revision<Long, Instruction>> entities = auditService.findRevisionByIdAndChangeKindNotIn(id,changekinds, pageable);
         return new ResponseEntity<>(assembler.toResource(entities), HttpStatus.OK);
     }
 }
