+++ src/main/java/no/nsd/qddt/domain/instrument/audit/InstrumentAuditService.java
@@ -13,7 +13,7 @@ import java.util.UUID;
 /**
  * @author Dag Østgulen Heradstveit
  */
-public interface InstrumentAuditService extends BaseServiceAudit<Instrument, UUID, Integer> {
+public interface InstrumentAuditService extends BaseServiceAudit<Instrument, UUID, Long> {
 
-    Page<Revision<Integer, Instrument>> findRevisionByIdAndChangeKindNotIn(UUID id, Collection<AbstractEntityAudit.ChangeKind> changeKinds, Pageable pageable);
+    Page<Revision<Long, Instrument>> findRevisionByIdAndChangeKindNotIn(UUID id, Collection<AbstractEntityAudit.ChangeKind> changeKinds, Pageable pageable);
 }
